requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-openai
chromadb>=0.5.0
sqlalchemy
pydantic
mcp
python-dotenv
langgraph-checkpoint-sqlite
arize
arize-otel
openinference-instrumentation-langchain
tavily-python
pandas
openpyxl
langchain_community
streamlit
nest-asyncio
pydantic-settings
langchain-huggingface
sentence-transformers
langchain_chroma